requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-magic-bin==0.4.14 ; sys_platform == 'win32'
python-magic==0.4.27 ; sys_platform != 'win32'
sqlparse==0.4.4
typing_extensions==4.11.0
